Use XPoint struct.  Reorder declarations per C99 style, latest possible.

+++ scr_insert_characters.h
@@ -1,7 +1,9 @@
 #ifndef SCR_INSERT_CHARACTERS_H
 #define SCR_INSERT_CHARACTERS_H
 
+#include <stdint.h>
+
 //  Insert count spaces from the current position.
-void scr_insert_characters(int count);
+void scr_insert_characters(int16_t count);
 
 #endif//!SCR_INSERT_CHARACTERS_H
